libraries/rust-speedup-showcase/requirements-python.txt
# Python (pure) versions for comparison
charset-normalizer>=3.3.0
packaging>=23.0
python-dateutil>=2.8.2
colorama>=0.4.6
tabulate>=0.9.0
humanize>=4.9.0
validators>=0.22.0

# Additional dependencies for benchmarking
numpy>=1.24.0
